Added score formatter

+++ Assets/Scripts/ShopButton.cs
@@ -6,7 +6,7 @@ public abstract class ShopButton : MonoBehaviour
     public Watermelon watermelon;
     public TMP_Text priceText;
     public GameObject decoration;
-    public int price;
+    public double price;
 
     public abstract void Buy();
 }
